requirements.txt
# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0

# Data processing & analysis
pandas>=2.0.0
numpy>=1.23.0

# HTTP & API
requests>=2.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Machine Learning
catboost>=1.1.0

# File I/O
pyarrow>=12.0.0

# Environment & config management
python-dotenv>=1.0.0

# Optional: scheduling (uncomment if needed)
# schedule>=1.2.0
# apache-airflow>=2.5.0  # for orchestration (install separately if needed)
